rearrange dependencies
