+++ frontend/src/app/document/document.module.ts
@@ -4,6 +4,7 @@ import { DocumentViewComponent } from '../document-view/document-view.component'
 import { DocumentPageComponent } from '../document-page/document-page.component';
 import { ImageViewModule } from '../image-view/image-view.module';
 import { SearchRelevanceComponent } from '../search';
+import { CorpusModule } from '../corpus-header/corpus.module';
 
 
 
@@ -14,6 +15,7 @@ import { SearchRelevanceComponent } from '../search';
         SearchRelevanceComponent,
     ],
     imports: [
+        CorpusModule,
         SharedModule,
         ImageViewModule,
     ], exports: [
